unicorns

+++ Vector3.h
@@ -5,43 +5,43 @@
 class Vector3
 {
 public:
-	double elements[3];
+	float elements[3];
 
-	Vector3(double x, double y, double z);
+	Vector3(float x, float y, float z);
 	Vector3();
 
-	double x();
-	double y();
-	double z();
+	float x();
+	float y();
+	float z();
 
 	Vector3 operator+(); //Unary operators
 	Vector3 operator-(); 
 
-	double operator[](int index) const; //Read member
-	double& operator[](int index); //To write to a member
+	float operator[](int index) const; //Read member
+	float& operator[](int index); //To write to a member
 
 	Vector3& operator +=(const Vector3& otherVector); //Vector with vector
 	Vector3& operator*=(const Vector3 otherVector);
 	Vector3& operator/=(const Vector3 otherVector);
 	Vector3& operator-=(const Vector3 otherVector);
 
-	Vector3& operator*=(const double scalar);
-	Vector3& operator/=(const double scalar);
+	Vector3& operator*=(const float scalar);
+	Vector3& operator/=(const float scalar);
 
 	Vector3 unit();
-	double length();
+	float length();
 
 };
 
 std::ostream& operator<<(std::ostream& stream, const Vector3& vec);
 
-double dot(const Vector3& first, const Vector3& second);
+float dot(const Vector3& first, const Vector3& second);
 
 Vector3 operator+(const Vector3& first, const Vector3& second);
 Vector3 operator-(const Vector3& first, const Vector3& second);
 Vector3 operator*(const Vector3& first, const Vector3& second); //Note to self, component wise multiplication. Not dot.
 Vector3 operator/(const Vector3& first, const Vector3& second);
 
-Vector3 operator/(const Vector3& vector, double divisor);
-Vector3 operator*(double divisor, const Vector3& vector);
+Vector3 operator/(const Vector3& vector, float divisor);
+Vector3 operator*(float divisor, const Vector3& vector);
 Vector3 cross(Vector3& vec1, Vector3& vec2);
